Continue updating tests to use new fixtures
Affects: https://www.notion.so/lyrasis/Refactor-circulation-tests-to-use-pytest-fixtures-4f2603de236746aeaab123dcf12aa445
