+++ tests/core_new/test_external_search.py
@@ -59,8 +59,7 @@ from core.model.classification import Subject
 from core.model.work import Work
 from core.problem_details import INVALID_INPUT
 from core.util.datetime_helpers import datetime_utc, from_timestamp
-from tests.core.utils import DBStatementCounter, PerfTimer
-from tests.fixtures.database import DatabaseTransactionFixture
+from tests.fixtures.database import DatabaseTransactionFixture, DBStatementCounter, PerfTimer
 from tests.fixtures.search import EndToEndSearchFixture, ExternalSearchFixture
 
 RESEARCH = Term(audience=Classifier.AUDIENCE_RESEARCH.lower())
